requirements.txt
# Audio Capture
sounddevice>=0.4.6
numpy>=1.24.0

# Transcription (Local Whisper)
openai-whisper
torch
torchaudio

# Claude API
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0

# Audio file handling
pydub>=0.25.0

# Speaker Diarization
pyannote.audio>=3.1.0

# Web Interface
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
qrcode>=7.4.0
pillow>=10.0.0
netifaces>=0.11.0

# Video Conferencing Support
psutil>=5.9.0
